+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+<<<<<<< Updated upstream
 ## 2.1.1
+=======
+## 2.2.1
+
+Increase scan device performance.
+
+## 2.2.0
+
+Switched from Bluetooth Thermal Printer library to Flutter Bluetooth Classic.
+
+## 2.1.1
+
+Updated some dependencies.
+
+## 2.1.0
+
+Switched from Bluetooth Thermal Printer library to Bluetooth Serial.
+
+## 2.0.3
+>>>>>>> Stashed changes
 
 Update Screenshot library to compatible with newer flutter version